backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
motor
python-dotenv
pydantic
pydantic-settings